+++ section_to_course/__init__.py
@@ -2,4 +2,4 @@
 Factors sections from Open edX courses into their own new course.
 """
 
-__version__ = '0.1.0'
+__version__ = '0.2.0'