+++ section_to_course/apps.py
@@ -3,6 +3,7 @@ section_to_course Django application initialization.
 """
 
 from django.apps import AppConfig
+from edx_django_utils.plugins.constants import PluginSettings, PluginURLs
 
 
 class SectionToCourseConfig(AppConfig):
@@ -12,4 +13,29 @@ class SectionToCourseConfig(AppConfig):
 
     name = 'section_to_course'
 
-    plugin_app = {}
+    plugin_app = {
+        # Configuration setting for Plugin URLs for this app.
+        PluginURLs.CONFIG: {
+            'cms.djangoapp': {
+                PluginURLs.NAMESPACE: 'section_to_course',
+                # The application namespace to provide to django's urls.include.
+                # Optional; Defaults to None.
+                PluginURLs.APP_NAME: 'section_to_course',
+
+                # The regex to provide to django's urls.url.
+                # Optional; Defaults to r''.
+                PluginURLs.REGEX: r'^section-to-course/api/',
+
+                # The python path (relative to this app) to the URLs module to be plugged into the project.
+                # Optional; Defaults to 'urls'.
+                PluginURLs.RELATIVE_PATH: 'api.urls',
+            }
+        },
+        PluginSettings.CONFIG: {
+            'cms.djangoapp': {
+                'common': {
+                    PluginSettings.RELATIVE_PATH: 'settings.common',
+                },
+            }
+        },
+    }