+++ section_to_course/management/commands/tests/test_selection_to_course.py
@@ -3,7 +3,7 @@ Tests the SectionToCourse management command
 """
 from io import StringIO
 
-from common.djangoapps.student.tests.factories import UserFactory
+from common.djangoapps.student.tests.factories import UserFactory  # pylint: disable=import-error
 from django.core.management import call_command
 
 try:
@@ -12,7 +12,7 @@ except ImportError:
     # This is no longer needed in Palm.
     from xmodule.modulestore.tests.factories import ItemFactory as BlockFactory, CourseFactory
 
-from xmodule.modulestore.tests.utils import MixedSplitTestCase
+from xmodule.modulestore.tests.utils import MixedSplitTestCase  # pylint: disable=import-error
 
 from section_to_course.models import SectionToCourseLink