+++ section_to_course/settings/common.py
@@ -1,0 +1,10 @@
+"""
+Tweaks the platform settings to enable support for section_to_course.
+"""
+
+
+def plugin_settings(settings):
+    """Add django_object_actions to the installed apps so that its templates are loaded."""
+    settings.INSTALLED_APPS += (
+        'django_object_actions',
+    )