+++ test_utils/test_settings.py
@@ -1,3 +1,0 @@
-"""Settings module for doc generation."""
-
-SECRET_KEY = 'bogus key'